+++ src/main/Java/models/Account.java
@@ -2,7 +2,7 @@ package models;
 
 public class Account {
     private String accountType;
-    private Integer accountId;
+    private Long accountId;
     private Double balance;
     private String status;
 }
